+++ authlete/dto/introspection_request.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -24,7 +24,10 @@ class IntrospectionRequest(Jsonable):
             'token':             str,
             'scopes':            str,  # list of str
             'subject':           str,
-            'clientCertificate': str
+            'clientCertificate': str,
+            'dpop':              str,
+            'htm':               str,
+            'htu':               str,
         }
 
         super().__init__(nameAndValues, nameAndTypes)